% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABneutral.R
\name{ABneutral}
\alias{ABneutral}
\title{Run Model with no selection (ABneutral)}
\usage{
ABneutral(pedigree.data, p0uu, eqp, eqp.weight, Nstarts, out.dir, out.name)
}
\arguments{
\item{pedigree.data}{Generation table name, you can find sample file in}

\item{p0uu}{Type of cytosine (CHH/CHG/CG)}

\item{eqp}{Filter value, based on posteriorMax}

\item{eqp.weight}{number of threads, default is 2.}

\item{Nstarts}{Nstarts better to set 50}

\item{out.dir}{outputdirectory}

\item{out.name}{name of file}
}
\value{
ABneutral data.
}
\description{
Run Model with no selection (ABneutral)
}
\examples{
## Get some toy data
file1 <- system.file("extdata/dm/","pedigree.csv", package="alphabeta")
pedigree <- as.matrix(read.table(file1,sep=",", header=TRUE, stringsAsFactors = FALSE))
p0uu_in <- 0.7435074
eqp.weight <- 1
Nstarts <- 5
output.data.dir <- paste0( getwd(),"/")
out.name <- "CG_global_estimates_ABneutral"
out1 <- ABneutral(pedigree.data = pedigree,
                  p0uu=p0uu_in,
                  eqp=p0uu_in,
                  eqp.weight=eqp.weight,
                  Nstarts=Nstarts,
                  out.dir=output.data.dir,
                  out.name="CG_global_estimates_ABneutral")

summary(out1)

}
