% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABselectMM.R
\name{ABselectMM}
\alias{ABselectMM}
\title{Run model with selection against spontaneous gain of methylation (ABselectMM)}
\usage{
ABselectMM(pedigree.data, p0uu, eqp, eqp.weight, Nstarts, out.dir,
  out.name)
}
\arguments{
\item{pedigree.data}{pedigree data.}

\item{p0uu}{initial proportion of unmethylated cytosines.}

\item{eqp}{equilibrium proportion of unmethylated cytosines.}

\item{eqp.weight}{nweight assigned to equilibrium function.}

\item{Nstarts}{iterations for non linear LSQ optimization.}

\item{out.dir}{output directory.}

\item{out.name}{output file name.}
}
\value{
ABselectMM  RDATA
}
\description{
Run model with selection against spontaneous gain of methylation (ABselectMM)
}
\examples{
## Get some toy data
file1 <- system.file("extdata/dm/","pedigree.csv", package="AlphaBeta")
pedigree <- as.matrix(read.table(file1, sep=",", header=TRUE, stringsAsFactors = FALSE))
p0uu_in <- 0.7435074
eqp.weight <- 1
Nstarts <- 4
output.data.dir <- paste0( getwd(),"/")
out.name <- "CG_global_estimates_ABselectMM"
out2 <- ABselectMM(pedigree.data = pedigree,
                  p0uu=p0uu_in,
                  eqp=p0uu_in,
                  eqp.weight=eqp.weight,
                  Nstarts=Nstarts,
                  out.dir=output.data.dir,
                  out.name=out.name)

summary(out2)

}
