% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABselectUU.R
\name{ABselectUU}
\alias{ABselectUU}
\title{Run model with selection against spontaneous loss of methylation (ABselectUU)}
\usage{
ABselectUU(pedigree.data, p0uu, eqp, eqp.weight, Nstarts, out.dir,
  out.name)
}
\arguments{
\item{pedigree.data}{pedigree data.}

\item{p0uu}{initial proportion of unmethylated cytosines.}

\item{eqp}{equilibrium proportion of unmethylated cytosines.}

\item{eqp.weight}{weight assigned to equilibrium function.}

\item{Nstarts}{iterations for non linear LSQ optimization.}

\item{out.dir}{output directory.}

\item{out.name}{output file name.}
}
\value{
ABselectMM  RData file.
}
\description{
This model assumes that heritable gains of cytosine methylation are under negative selection.
}
\examples{
## Get some toy data
inFile <- system.file("extdata/dm/","pedigree.csv", package="AlphaBeta")
pedigree <- as.matrix(read.table(inFile, sep=",", header=TRUE, stringsAsFactors = FALSE))
p0uu_in <- 0.7435074
eqp.weight <- 1
Nstarts <- 2
output.data.dir <- paste0( getwd(),"/")
out.name <- "CG_global_estimates_ABselectUU"
out3 <- ABselectUU(pedigree.data = pedigree,
                  p0uu=p0uu_in,
                  eqp=p0uu_in,
                  eqp.weight=eqp.weight,
                  Nstarts=Nstarts,
                  out.dir=output.data.dir,
                  out.name=out.name)

summary(out3)

}
