% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABselectUUSOMA.R
\name{ABselectUUSOMA}
\alias{ABselectUUSOMA}
\title{Model with selection against spontaneous loss of methylation (outselectUU)}
\usage{
ABselectUUSOMA(pedigree.data, p0uu, eqp, eqp.weight, Nstarts, out.dir,
  out.name)
}
\arguments{
\item{pedigree.data}{pedigree data.}

\item{p0uu}{initial proportion of unmethylated cytosines.}

\item{eqp}{equilibrium proportion of unmethylated cytosines.}

\item{eqp.weight}{weight assigned to equilibrium function.}

\item{Nstarts}{iterations for non linear LSQ optimization.}

\item{out.dir}{output directory.}

\item{out.name}{output file name.}
}
\value{
ABneutralSoma data.
}
\description{
Model with selection against spontaneous loss of methylation (outselectUU)
}
\examples{
## Get some toy data
file1 <- system.file("extdata/soma/","pedigreeSoma.csv", package="AlphaBeta")
pedigree <- as.matrix(read.table(file1, sep=",", header=TRUE, stringsAsFactors = FALSE))
p0uu_in <- 0.54755
eqp.weight <- 0.001
Nstarts <- 2
output.data.dir <- paste0( getwd(),"/")
out.name <- "ABselectUUSOMA_CG_estimates"
out <- ABselectUUSOMA(pedigree.data = pedigree,
                  p0uu=p0uu_in,
                  eqp=p0uu_in,
                  eqp.weight=eqp.weight,
                  Nstarts=Nstarts,
                  out.dir=output.data.dir,
                  out.name=out.name)

summary(out)

}
