% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BOOTmodel.R
\name{BOOTmodel}
\alias{BOOTmodel}
\title{Bootstrap analysis with the best model}
\usage{
BOOTmodel(pedigree.data, Nboot, out.dir, out.name)
}
\arguments{
\item{pedigree.data}{pedigree data.}

\item{Nboot}{number of boot.}

\item{out.dir}{output directory.}

\item{out.name}{output file name.}
}
\value{
bootstrap result.
}
\description{
Bootstrap analysis with the best model
}
\examples{
## Get some toy data
inFile <- system.file("extdata/models/","ABneutral_CG_global_estimates.Rdata", package="AlphaBeta")
Nboot <- 4
out.name <-"Boot_CG_global_estimates_ABneutral"
Bout <- BOOTmodel(pedigree.data=inFile,
                Nboot=Nboot,
                out.dir=getwd(),
                out.name=out.name)
summary(Bout)
}
