% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertDMATRIX.R
\name{convertDMATRIX}
\alias{convertDMATRIX}
\title{convertDMATRIX}
\usage{
convertDMATRIX(sample.info, branch.points, dmatrix, design)
}
\arguments{
\item{sample.info}{input file containing information on generation times and pedigree lineages}

\item{branch.points}{input file containing lineage branch points}

\item{dmatrix}{input file containing 5mC divergence values for each sample pair}

\item{design}{"sibling" or "direct"}
}
\value{
pedigree
}
\description{
calculate divergence times of the pedigree
}
\examples{
## Get some toy data
file1 <- system.file("extdata/dm/","sampleInfo.csv", package="AlphaBeta")
file2<-system.file("extdata/dm/","branchPoints.csv", package="AlphaBeta")
file3<-system.file("extdata/dm/","d-matrix-CG.csv", package="AlphaBeta")
sample.info <-read.table(file1,sep="\\t", header=TRUE, stringsAsFactors = FALSE)
branch.points <-read.table(file2,sep="\\t", header=TRUE, stringsAsFactors = FALSE)
dmatrix <-read.table(file3,sep="\\t", header=TRUE, stringsAsFactors = FALSE)
pedigree <- convertDMATRIX(sample.info=sample.info,
 branch.points=branch.points,dmatrix=dmatrix,design="sibling")
head(pedigree)
}
