% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D-matrices.R
\name{dMatrices}
\alias{dMatrices}
\title{Constructing D-Matrices}
\usage{
dMatrices(directory, cytosines, posteriorMaxFilter, genTable)
}
\arguments{
\item{directory}{Path to the data-source directory}

\item{cytosines}{Type of cytosines (CHH/CHG/CG)}

\item{posteriorMaxFilter}{Filter value, based on posteriorMax ex: >= 0.95 or 0.99}

\item{genTable}{Generation table name, you can find sample file in "extdata" generations.fn}
}
\value{
generating divergence matrices file.
}
\description{
Estimating epimutation rates from high-throughput DNA methylation data
}
\examples{
dMatrices("data-source directory", "CHH", 0.99, "table.fn")
generation table name should be in this format:
system.file("extdata","generations.fn", package = "alphabeta")
}
