% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcMethlvl.R
\name{rc.meth.lvl}
\alias{rc.meth.lvl}
\title{Calculating rc.Meth.lvl}
\usage{
rc.meth.lvl(nodelist, cytosine, posteriorMaxFilter)
}
\arguments{
\item{nodelist}{List of samples, you can find sample file in
"extdata" called "nodelist.fn"}

\item{cytosine}{Type of cytosine (CHH/CHG/CG)}

\item{posteriorMaxFilter}{Filter value, based on posteriorMax}
}
\value{
rc meth lvl.
}
\description{
Estimating epimutation rates from high-throughput DNA methylation data
}
\examples{
## Get some toy data
file <- system.file("extdata/dm/","tmp_nodelist.fn", package="AlphaBeta")
rc.meth.lvl(file, "CG", 0.99)
}
