% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABnull.R
\name{ABnull}
\alias{ABnull}
\title{Run model that considers no accumulation of epimutations (ABnull)}
\usage{
ABnull(pedigree.data, out.dir, out.name)
}
\arguments{
\item{pedigree.data}{Generation table name, you can find sample file in}

\item{out.dir}{outputdirectory}

\item{out.name}{name of file}
}
\value{
ABnull data.
}
\description{
Run model that considers no accumulation of epimutations (ABnull)
}
\examples{
# Get some toy data
file1 <- system.file("extdata/dm/","pedigree.csv", package="alphabeta")
pedigree <- as.matrix(read.table(file1,sep=",", header=TRUE, stringsAsFactors = FALSE))
output.data.dir <- paste0( getwd(),"/")
out.name <- "CG_global_estimates_ABnull"
out1 <- ABnull(pedigree.data = pedigree,
                  out.dir=output.data.dir,
                  out.name=out.name)

summary(out1)

}
