% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BOOTmodel.R
\name{BOOTmodel}
\alias{BOOTmodel}
\title{Bootstrap analysis with the best model}
\usage{
BOOTmodel(pedigree.data, Nboot, out.dir, out.name)
}
\arguments{
\item{pedigree.data}{pedigree data.}

\item{Nboot}{number of boot.}

\item{out.dir}{output directory.}

\item{out.name}{output file name.}
}
\value{
outfinal compared data.
}
\description{
Bootstrap analysis with the best model
}
\examples{
## Get some toy data
file1 <- system.file("extdata/models/","CG_global_estimates_ABneutral.Rdata", package="AlphaBeta")
Nboot <- 4
out.dir <- paste0( getwd(),"/")
out.name <-"Boot_CG_global_estimates_ABneutral"
Bout <- BOOTmodel(pedigree.data=file1,
                Nboot=Nboot,
                out.dir=out.dir,
                out.name=out.name)
}
