% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D-matrices.R
\name{dMatrices}
\alias{dMatrices}
\title{Constructing D-Matrices}
\usage{
dMatrices(directory, cytosines, posteriorMaxFilter)
}
\arguments{
\item{directory}{Path to the data-source directory}

\item{cytosines}{Type of cytosines (CHG or CG or CHH)}

\item{posteriorMaxFilter}{Filter value, based on posteriorMax ex: >= 0.95 or 0.99}
}
\value{
A matrix of the infile
}
\description{
This function loads a file as a matrix. It assumes that the first column
contains the rownames and the subsequent columns are the sample identifiers.
Any rows with duplicated row names will be dropped with the first one being
kepted.
}
