% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D-matrices.R
\name{dMatrices}
\alias{dMatrices}
\title{Constructing D-Matrices}
\usage{
dMatrices(genTable, cytosines, posteriorMaxFilter)
}
\arguments{
\item{genTable}{Generation table name, you can find sample file in
"extdata" called "generations.fn"}

\item{cytosines}{Type of cytosines (CHH/CHG/CG)}

\item{posteriorMaxFilter}{Filter value, based on posteriorMax
ex: >= 0.95 or 0.99}
}
\value{
generating divergence matrices file.
}
\description{
Estimating epimutation rates from high-throughput DNA methylation data
}
\examples{
## Get some toy data
file <- system.file("extdata","generations.fn", package="alphabeta")

dMatrices(file, "CG", 0.99)
}
