% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePHYLO.R
\name{makePHYLO}
\alias{makePHYLO}
\title{Generate pedigree}
\usage{
makePHYLO(tall, pedigree, sample.info)
}
\arguments{
\item{tall}{pedigree data.}

\item{pedigree}{dmatrix.}

\item{sample.info}{sample.info}
}
\value{
pedigree data.
}
\description{
Generate pedigree
}
\examples{
## Get some toy data
file1 <- system.file("extdata/soma/","D-matrix_CG.csv", package="alphabeta")
file2 <- system.file("extdata/soma/","sample_info.csv", package="alphabeta")
d.matrix <- read.table(file1, sep="\\t", header=TRUE, stringsAsFactors = FALSE)
sample.info <- read.table(file2, sep="\\t", header=TRUE, stringsAsFactors = FALSE)
out <- makePHYLO(tall=330, pedigree = d.matrix, sample.info = sample.info)

}
