% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePHYLO.R
\name{makePHYLO}
\alias{makePHYLO}
\title{Generating tree pedigree data}
\usage{
makePHYLO(tall, pedigree, sample.info)
}
\arguments{
\item{tall}{total age of the tree.}

\item{pedigree}{dmatrix file.}

\item{sample.info}{sample info file.}
}
\value{
tree pedigree data file.
}
\description{
Generating tree pedigree data
}
\examples{
## Get some toy data
file1 <- system.file("extdata/soma/","AB-dMatrix-CG-0.99.csv", package="AlphaBeta")
file2 <- system.file("extdata/soma/","sampleInfo.csv", package="AlphaBeta")
d.matrix <- read.table(file1, sep="\\t", header=TRUE, stringsAsFactors = FALSE)
sample.info <- read.table(file2, sep="\\t", header=TRUE, stringsAsFactors = FALSE)
# in our case, the total age of tree is 330
out <- makePHYLO(tall=330, pedigree = d.matrix, sample.info = sample.info)

}
